// src/app/page/master/penghasilan-tidak-kena-pajak/penghasilan-tidak-kena-pajak.interface.ts
import { EventEmitter } from '@angular/core';

export interface PenghasilanTidakKenaPajak {
    kode?: any;
    kdStatusPerkawinan?: number;
    statusPTKP?: string;
    deskripsi?: string;
    qtyAnak?: string;
    totalHargaPTKP?: string;
    reportDisplay?: string;
    kodeExternal?: string;
    namaExternal?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}